ARM: dts: Set GPU to use DPLL_GPU as clock source
In the current clock configuration, GPU derives its clocks
from DPLL_CORE H14 output called CORE_GPU_CLK. The multipliers,
dividers and post-dividers are set in the u-boot according to the
desired clock rate.

However, if it is required to set GPU to different OPPs based on
u-boot config, kernel clock framework requires
a. either changing the H14 post-divider, which is an impossible integer
   for GPU OPP clock recommendation (425.6 MHz), or
b. affect the clock setting for other clocks of DPLL_CORE.

It is, therefore, desirable to set GPU to use DPLL_GPU to derive its
clocks by using the DT standard properties "assigned-clocks" and
"assigned-clock-parents".

This patch also sets the initial values for the DPLL_GPU outputs. These values
are chosen based on the OPP_NOM values defined as per recommendation from
design team. The DPLL locked frequency is kept at 1277 MHz, so that the value
for the divider clock, dpll_gpu_m2_ck, can be set to 425.67 MHz for OPP_NOM.

Signed-off-by: Subhajit Paul <subhajit_paul@ti.com>