remoteproc: pruss: fix pruss_get() to support deferred probe in clients
Knowing the error cause in pruss_get() is required to implement deferred
probe in the prueth driver and other PRUSS client drivers: deferred probe
must take place only when the pruss device is not yet available, not when
it is never going to be available. Fix up the pruss_get() function to
return an appropriate error-scenario specific ERR_PTR instead of a NULL
value on all errors.

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>
Acked-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>