soc: ti: wkup_m3_ipc: Fix wkup_m3_ipc debugfs entry
There are several issues with the wkup_m3_ipc debugfs entry and this
patch fixes them all:

In an attempt to prevent the enable_late_halt debugfs entry from using
an invalid wkup_m3_ipc pointer its creation was moved until after the
wkup_m3 had booted. Unfortunately it was placed in the wkup_m3 irq
handler leading to a bug.

Waiting until the m3 is booted is unneccessary as the data structures
configured by the enable_late_halt sysfs entry are created during probe,
so call wkup_m3_ipc_dbg_init during probe instead of interrupt context
which causes "sleeping function called from invalid context
triggered by wkup_m3_txev_handler" bug.

In addition to that the wkup_m3_ipc_dbg_init call is enclosed in
CONFIG_DEBUG_FS block but will cause undefined function error if that
option is not selected, so add a stub for case where !CONFIG_DEBUG_FS.

Finally, make sure we remove the debugfs entry when we remove the driver
so there are no issues when we try to prove again, and so that there is
not a floating, invalid, debugfs entry when the driver is unloaded.

Fixes: f5714f575f3b ("soc: ti: wkup_m3_ipc: Fix enable_late_halt sysfs entry")
Signed-off-by: Dave Gerlach <d-gerlach@ti.com>