arm: dtsi: am57xx-evm-common: mt9t11x set vsync active low
The MT9T111 sensor does not actually define and use
Vsync and Hsync signals instead it defines line valid (LV) and
frame valid (FV) instead. FV is high during the entire valid video
frame. This is the opposite meaning of what VIP interpret Vsync to
be. This inversion introduces latency in frame detection within the
VIP hardware. Since FV is the opposite of Vsync then we better
let VIP know that Vsync is active low instead of high.

Signed-off-by: Benoit Parrot <bparrot@ti.com>