ARM: dts: am335x-evmsk: prevent wilink8 from blocking suspend
mmc2 used for wl12xx was missing the keep-power-in suspend
parameter. As a result the board couldn't reach suspend state.

In addition Wilink8 supports wowlan (wake on wlan)
This flag has to be there for all boards so this feature can work.
When using wowlan the wilink8 chip (internal firmware) has to
stay powered-on during suspend and keep running so it can
trigger a wakeup event to the host when a "magic" packet is
received.

Signed-off-by: Eyal Reizer <eyalr@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>