ARM: dts: beagle-x15-common: Add watchdog timer for DSP2
The watchdog timer information has been added to the DSP2 remote
processor device node on all the AM57xx BeagleBoard-X15 boards.
The watchdog timers for other processors have already been added.
This data is identical to the timers used on the DRA7 EVM board
to maintain firmware compatibility between both the SoC and board
variants. The watchdog functionality will use GPTimer 13 for the
DSP processor core.

The MPU-side drivers will use this data to initialize the watchdog
timer, and listen for any watchdog triggers. The BIOS-side code on
DSP2 needs to configure/refresh this timer properly to not throw a
watchdog error.

This timer can be changed or removed as per the system integration
needs, alongside appropriate equivalent changes on the firmware side.

Signed-off-by: Suman Anna <s-anna@ti.com>