ARM: dts: am572x-idk: Add watchdog timer for DSP2
The watchdog timer information has been added to the DSP2 remote
processor device node on the AM572x IDK boards. The watchdog timers
for other processors have already been added. This data is identical
to the timer used on the AM57xx EVM board board to maintain firmware
compatibility between both the SoC and board variants. The watchdog
functionality will use GPTimer 13 for the DSP processor core.

The MPU-side drivers will use this data to initialize the watchdog
timer, and listen for any watchdog triggers. The BIOS-side code on
DSP2 needs to configure/refresh this timer properly to not throw a
watchdog error.

This timer can be changed or removed as per the system integration
needs, alongside appropriate equivalent changes on the firmware side.

Signed-off-by: Suman Anna <s-anna@ti.com>