remoteproc/keystone: create sysfs entries for memories
The keystone-remoteproc driver is enhanced to provide sysfs entries
for each DSP remoteproc processor to allow the userspace to read the
address and size of the various DSP internal RAM memories that are
exposed to userspace through a corresponding misc device. This sysfs
logic allows the userspace-based Multi Proc Manager (MPM) stack to
not rely on procfs-based DT parsing for looking up the memories.

Each supported memory region is represented by its own directory and
presents two sysfs files 'addr' and 'size', and are created under the
respective dspX misc device. The directories can be accessed under the
/sys/class/misc/dspX/ path, where X is the DSP number (indexed from 0).

Signed-off-by: Suman Anna <s-anna@ti.com>