net: prueth: Sync header with firmware rev 03.01.04
The DRAM start offset and PORT_QUEUE_DESC_OFFSET has been changed
in firmware release 03.01.04

This was causing ingress packets in higher priority queue to be dropped.
e.g. (VLAN packets with priority 4 to 7 which are put by firmware
in higher priority queue i.e. Queue 1 for port 1 and Queue 2 for port 2).

Signed-off-by: Roger Quadros <rogerq@ti.com>
[nsekhar@ti.com: convert tab after #define to space, move hex to lowercase]
Signed-off-by: Sekhar Nori <nsekhar@ti.com>