drm/omap: use DRM_ROTATE_* instead of OMAP_DSS_ROT_*
commit a9f673b250ac ("drm/omap: use DRM_ROTATE_* instead of OMAP_DSS_ROT_*") branch ti-linux-4.9.y

At the moment the dispc driver uses a custom enum for rotation. Change
it to use the DRM's DRM_ROTATE_*.

Note that mirroring is at the moment handled as a separate boolean in
the dispc driver, so we only use the DRM_ROTATE_* values.

Note, DSS HW uses clockwise rotation, DRM counter-clockwise.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>