HACK: input: touchscreen: edt-ft5x06: fix driver autoprobing
Because with OF we can pass more specific
compatible flags (such as edt-ft5306) instead
of generic edt-ft5x06, when i2c-core's
of_i2c_register_devices() tries to request_module(),
it'll request it with a non-existent specific module
alias.

In order to have this driver autoprobing again, we
just need to add missing MODULE_ALIAS() entries to
edt-ft5x06 driver.

Thanks to Dmitry for noticing that it actually should
autoprobe even with of_device_id.

This patch is marked as a "HACK" since this is not
acceptable upstream and a solution is still being
debated here: https://lkml.org/lkml/2015/9/9/422

Signed-off-by: Felipe Balbi <balbi@ti.com>
[fcooper@ti.com: tweaked to add support FT5506]
Signed-off-by: Franklin S Cooper Jr <fcooper@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>