net: prueth: fix probe crash on sram memory allocation failure
gen_pool_alloc() returns 0 if requested size is bigger than available in
sram pool. Then prueth->mem[PRUETH_MEM_OCMC].va is NULL, but that is not
detected.

A bit later, another init fails and gen_pool_free() is called to free
the sram memory, which is not allocated. Leading to:
[    3.338790] Kernel BUG at c03d58dc [verbose debug info unavailable]
[    3.378568] PC is at gen_pool_free+0xd4/0xd8
...
[    3.623594] Backtrace:
[    3.626186] [<c03d5808>] (gen_pool_free) from [<c0579684>] (prueth_probe+0x4e8/0x116c)

This patch handles gen_pool_alloc() returning 0 and exits probe nicely.

Acked-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Pascal Eberhard <pascal.eberhard@gmail.com>
[nsekhar@ti.com: use simpler check than IS_ERR_OR_NULL(), fix return value
		 and minor commit message adjustments]
Signed-off-by: Sekhar Nori <nsekhar@ti.com>