Revert "TEMP: iommu/omap: fix range for cache flush operations"
This reverts commit 4424597becb984d6e38ee28cf642031bad4b4a1c.

Undo the changes added to the OMAP IOMMU driver in commit 4424597becb9
("TEMP: iommu/omap: fix range for cache flush operations"), which fixed
an issue with some of the pgd and pte entries not getting flushed
completely due to incorrect ranges being passed into the cache flush
operations. This was needed with the changes added in commit 59116088bdb3
("HACK: iommu/omap: flush page table entries from L2 cache") to flush
the page table entries from the L2 cache.

This reverts the temporary commit in preparation of a proper fix with the
DMA Streaming API. The OMAP IOMMU driver will use the DMA Streaming API
to achieve cache functionality by flushing the page table/directory table
entries from the CPU caches. This operation must be done so that the
contents of the cache is written back to main memory and the device is
then able to read it.

Signed-off-by: Josue Albarran <j-albarran@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>