Revert "HACK: iommu/omap: flush page table entries from L2 cache"
This reverts commit 59116088bdb373c13a97f5a4c7ee1418d7095d5d.

Undo the changes added to the OMAP IOMMU driver in commit 59116088bdb3
("HACK: iommu/omap: flush page table entries from L2 cache"), which added
the proper outer cache operations to flush the page table entries from
L2 cache. L2 cache was not being flushed before and this resulted in
MMU faults on the DSP and IPU processors on OMAP4.

This reverts the temporary commit in preparation of a proper fix with the
DMA Streaming API. The OMAP IOMMU driver will use the DMA Streaming API
to achieve cache functionality by flushing the page table/directory table
entries from the CPU caches. This operation must be done so that the
contents of the cache is written back to main memory and the device is
then able to read it.

Signed-off-by: Josue Albarran <j-albarran@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>