ARM: OMAP2+: sleep33xx/sleep43xx: Shuffle EMIF configuration
Currently we are saving the context of the EMIF before programming the
EMIF to enter self-refresh mode. In order to ensure that we can properly
resume, make sure that we program the EMIF to enter self-refresh before
we save the context of the EMIF so that the actual state of the EMIF is
what gets restored on resume. This allows us to properly clear the
self-refresh state if the EMIF is set for that after restoring the
context and guaranteeing proper exit of low power states.

Otherwise the EMIF will always be programmed on resume as not being in
self-refresh and it is critical that the EMIF is brought back up
configured for self-refresh and then transitioned into normal operation,
otherwise self-refresh exit may fail and hang the board because the
contents of memory is lost.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>