ARM: OMAP2+: pm33xx-core: Do not set pwr_mode in omap2_gpio_prepare_for_idle
In omap2_gpio_prepare_for_idle there is a pwr_mode argument for
describing whether or not the system is going to off mode or not.
However, this specific flag seems to be for OMAP2420 Errata item
1.101, and there does not seem to be any indication that this is needed
for AM437x.

With this flag set gpio0 wake from DeepSleep no longer works after an
off-mode transtion. emoving the flag not only causes no observable
difference in behavior but also allows gpio09 wake to continue working
after a transition.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>