net/rpmsg: fix kernel crash for Rx messages on probed devices
The rpmsg proto devices that are probed do create an embedded rpmsg
endpoint for receiving messages, but are not really designed to
process any such received messages. The endpoint's callback field
was repurposed to store and maintain a list of the rpmsg sockets
bound to this device. This list is used to mark the linked sockets
with appropriate status for achieving error recovery functionality.
Any messages sent to this endpoint are reusing the function designed
to receive messages for sockets and results in a kernel crash due
to incorrect interpretation of the endpoint's private data field.

Fix this by reusing/renaming the current callback function to use
strictly for endpoints associated with userspace create rpmsg
sockets, and a different callback function to print a warning
trace about unexpected Rx message.

Reported-by: Sam Nelson <samnelson@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>