net/rpmsg: fix kernel crash for Rx messages on published channels
The rpmsg proto channel devices that are published from the remote
processor side also have an embedded rpmsg endpoint for receiving
messages, but are not really designed to process any such received
messages. The endpoint's callback field was repurposed to store and
maintain a list of the rpmsg sockets connected to this device. This
list is used to mark these linked sockets with appropriate status
for achieving error recovery functionality. Any messages sent to
this endpoint results in a kernel crash due to incorrect typecast
of the endpoint's private data field.

Fix this by checking for the range of the rpmsg device's src addr,
and limit the callback function to process messages only for rpmsg
channel devices associated with userspace created rpmsg sockets.

Signed-off-by: Suman Anna <s-anna@ti.com>