rpmsg: rpc: fix suspicious rcu_dereference_check() usage
The rpmsg rpc code uses fcheck() to verify that the passed-in
dma-buf file descriptors during buffer pre-registration or
unregistration are valid. The caller must hold a rcu though
as part of the lock-free semantics of the file descriptor
management. Add the appropriate rcu locking to fix the following
warning thrown when CONFIG_PROVE_RCU is enabled:

 ===============================
 [ INFO: suspicious RCU usage. ]
 4.9.54-02664-g557222289973 #410 Not tainted
 -------------------------------
 ./include/linux/fdtable.h:93 suspicious rcu_dereference_check() usage!

 other info that might help us debug this:

 rcu_scheduler_active = 2, debug_locks = 1
 no locks held by multiqueue0:src/1066.

 stack backtrace:
 CPU: 0 PID: 1066 Comm: multiqueue0:src Not tainted 4.9.54-02664-g557222289973 #410
 Hardware name: Generic DRA74X (Flattened Device Tree)
 [<c011144c>] (unwind_backtrace) from [<c010c4f8>] (show_stack+0x10/0x14)
 [<c010c4f8>] (show_stack) from [<c0418fc0>] (dump_stack+0x98/0xc4)
 [<c0418fc0>] (dump_stack) from [<bf2e4e70>] (rppc_ioctl+0x514/0x730 [rpmsg_rpc])
 [<bf2e4e70>] (rppc_ioctl [rpmsg_rpc]) from [<c026f158>] (do_vfs_ioctl+0xa0/0xa10)
 [<c026f158>] (do_vfs_ioctl) from [<c026fafc>] (SyS_ioctl+0x34/0x5c)
 [<c026fafc>] (SyS_ioctl) from [<c0108240>] (ret_fast_syscall+0x0/0x1c)
Redistribute latency...

The fcheck() calls can perhaps be removed completely and relying on
the implicit checking during the dma_buf_get() calls, but is retained
to provide a preliminary check of the file descriptors.

Fixes: 06302fc65e3d ("rpmsg: rpc: introduce a new rpmsg_rpc driver")
Signed-off-by: Suman Anna <s-anna@ti.com>