rpmsg: rpc: fix suspicious rcu_dereference_check() usage
The rpmsg rpc code uses fcheck() to verify that the passed-in
dma-buf file descriptors during buffer pre-registration or
unregistration are valid. The caller must hold a rcu though
as part of the lock-free semantics of the file descriptor
management. Add the appropriate rcu locking to fix the following
warning thrown when CONFIG_PROVE_RCU is enabled:

 ===============================
 [ INFO: suspicious RCU usage. ]
 4.4.91-02991-gb242ddd5eda5 #275 Not tainted
 -------------------------------
 include/linux/fdtable.h:93 suspicious rcu_dereference_check() usage!

 other info that might help us debug this:

 rcu_scheduler_active = 1, debug_locks = 1
 no locks held by multiqueue0:src/1104.

 stack backtrace:
 CPU: 0 PID: 1104 Comm: multiqueue0:src Not tainted 4.4.91-02991-gb242ddd5eda5 #275
 Hardware name: Generic DRA74X (Flattened Device Tree)
 [<c0019054>] (unwind_backtrace) from [<c00147d4>] (show_stack+0x10/0x14)
 [<c00147d4>] (show_stack) from [<c02fbd48>] (dump_stack+0x98/0xc4)
 [<c02fbd48>] (dump_stack) from [<bf00d0d4>] (rppc_ioctl+0x3f0/0x6fc [rpmsg_rpc])
 [<bf00d0d4>] (rppc_ioctl [rpmsg_rpc]) from [<c015d57c>] (do_vfs_ioctl+0x47c/0x7a8)
 [<c015d57c>] (do_vfs_ioctl) from [<c015d8dc>] (SyS_ioctl+0x34/0x5c)
 [<c015d8dc>] (SyS_ioctl) from [<c0010540>] (ret_fast_syscall+0x0/0x1c)

The fcheck() calls can perhaps be removed completely and relying on
the implicit checking during the dma_buf_get() calls, but is retained
to provide a preliminary check of the file descriptors.

Fixes: 71711c3509d8 ("rpmsg: rpc: introduce a new rpmsg_rpc driver")
Signed-off-by: Suman Anna <s-anna@ti.com>