ti_config_fragments: defconfig_builder: Fix working directory
Fix the TI_WORKING_PATH if defconfig builder is run from
the ti_config_fragments directory or from a top level directory.

If defconfig_builder is run from the ti_config_fragments directory
then it could not find the map file in that directory.

This fixes that issue by grepping for ti_config_fragments in the
WORKING_PATH

Signed-off-by: Dan Murphy <dmurphy@ti.com>