ARM: dts: am437x-gp-evm: Remove clashing pinmux entries from unused pins
There are some pinmux entries in am437x-gp-evm which are entered in
unused pins but used by emmc which is disabled by default. The process
of enabling emmc requires the removal of these entries which is not
mentioned in the .dts comments. It is more suitable to not have these
entries than remove entries every time emmc is enabled.

Therefore, remove clashing pinmux entries from unused pins.

Signed-off-by: Faiz Abbas <faiz_abbas@ti.com>
Acked-by: Dave Gerlach <d-gerlach@ti.com>