clk: ti: divider: Update context save and restore
Commit 5c420dac7d46 ("clk: ti: divider: convert TI divider clock to use
its own data representation") changed how divider clocks are represented
in the driver, but because this was an upstream patch it did not update
the context save and restore functions needed for RTC+DDR mode.

This patch converts clk_divider_save_context and
clk_divider_restore_context to use the new struct clk_omap_divider
representation of divider clocks and adds the necessary "context" member
to store the context.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>