ARM: dts: am57xx-idk: Move spi-max-frequency out of common
Currently, am574x-idk.dts overrides the spi-max-frequency property of QSPI
controller and flash node that is specified in the common dtsi file.
This is bit confusing if one sees common dtsi file but fails to see the
board specific dts. So, lets not override this property but instead
populate in individual board dts files.

Signed-off-by: Vignesh R <vigneshr@ti.com>