PCI: dra7xx: Clear IRQSTATUS_MSI as soon as its read
If IRQSTATUS_MSI bit is cleared after handling interrupts,
it is possible that an interrupt comes in just after the specific
MSI vector is cleared but before IRQSTATUS_MSI is cleared. This
interrupt will not trigger a call to the handler and will be
missed.
To prevent this race, clear IRQSTATUS_MSI as soon as it is
read.

Signed-off-by: Faiz Abbas <faiz_abbas@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>