net/rpmsg: fix error status marking of new Rx sockets during recovery
The rpmsg proto driver creates a rpmsg channel device per remote
processor for each MessageQ object through the socket's bind()
call (a Rx socket) for receiving messages. These rpmsg channel
devices do not have a designated dst address upon creation, and
a dst address is automatically assigned the first time a message
is received on the respective channel. An error status is set on
all the bound (Rx) sockets during the auto-removal of the rpmsg
devices during a remote processor recovery, but this logic is
skipped for some devices with a dst address of RPMSG_ADDR_ANY.
This causes some Rx sockets to be not marked properly if an error
recovery happens before they receive a message, and this may result
in a kernel panic during the cleanup of the corresponding socket
upon the termination of the application.

Fix this by adjusting the logic in rpmsg_proto_remove() so that
all Rx sockets are marked properly irrespective of whether they
received a message or not. An additional check is also added
in the bind implementation to check for valid range of socket
addresses since the logic in the probe and remove functions
rely on this range to distinguish userspace created rpmsg
channel devices and those published from the remote processor.

Fixes: e5a9a772354e ("net/rpmsg: set error status on open sockets after recovery")
Reported-by: Angela Stegmaier <angelabaker@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>