remoteproc/pru: use macros and types from omap mailbox
The actual payload size for OMAP Mailboxes is 32-bits, and the type
to use with the OMAP Mailbox driver is mbox_msg_t. This type can vary
in size (32-bits or 64-bits) depending on the platform the driver is
being built for. Update the PRU remoteproc driver to use the types and
macros provided by the OMAP Mailbox driver to always convert safely
between the mailbox payload size and the pointers the data is exchanged
between the OMAP Mailbox implementation driver and its client drivers.

The OMAP Mailbox driver is built only for OMAP2+ and K3 builds. The PRU
remoteproc driver uses the OMAP mailbox types and macros unconditionally
though and do not require the OMAP Mailbox driver to be enabled. So,
define the mbox_type_t additionally for ARCH_KEYSTONE as well so that
there are no build errors seen with PRU remoteproc driver on Keystone 2
platforms.

Signed-off-by: Suman Anna <s-anna@ti.com>