dt-bindings: remoteproc: Add PRUSS bindings
This patch adds the bindings for the Programmable Real-Time Unit and
Industrial Communication Subsystem (PRU-ICSS) present on various
TI SoCs. The IP is present on multiple TI SoC architecture families
including the OMAP architecture SoCs such as AM33xx, AM437x and
AM57xx; and on a Keystone 2 architecture based 66AK2G SoC. It is
also present on the Davinci based OMAPL138 SoCs as well (not covered
for now). Details have been added to include bindings for various
core sub-modules like the PRU Cores, the PRUSS Interrupt Controller,
and other sub-modules used for Industrial Communication purposes,
covering the MDIO, MII_RT and the IEP sub-modules. The binding
mostly uses standard DT properties.

The binding status is also marked as unstable as the binding does
not cover all the sub-modules that are present within the PRU-ICSS,
and is also subject to improvements and changes upstream.

Signed-off-by: Suman Anna <s-anna@ti.com>