regulator: lp873x: Add support to hook regulators to EN pin
The EN signal coming in to the PMIC is actually used to poweroff
the system completely. There is no other mechanism for doing this
in lp873x based systems. If the EN bit is down for a specific regulator,
it won't go down when attempting to poweroff the system,
otherwise the state of a regulator is the result of (EN_PIN | EN_BIT).

Support hooking the regulators to EN Pin.

Signed-off-by: Keerthy <j-keerthy@ti.com>