media: ti-vpe: vip: Set bus_info to show VIP resource used
The bus_info property of the CAPS structure returned during a
VIDIOC_QUERYCAP was static. Instead we should populate it with the
actual detailed bus information corresponding to the actual physical VIP
port. I.E we should identify which VIN pins and stream number the video
node correspond to.

As video node creation is somewhat random and varies from boot to boot
cycle this is the only sure way to be able to identify a VIP
port/stream.

Signed-off-by: Benoit Parrot <bparrot@ti.com>