drm/tilcdc: rename tfp410 driver's name
commit 303110dfa3d9fd4ebf8b95ff94467a2ab177d02c ti-linux-4.9.y

The name field in the platform_driver for both omapdss's tfp410 and
tilcdc's tfp410 is set to the same value "tfp410", which causes a name
clash when both drivers are enabled in the kernel.

Afaics, the driver name is not used for much, other than in sysfs.

This patch renames tilcdc's tfp410 driver to "tilcdc-tfp410".

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>