ti_config_fragments: rpmsg: Add RPMsg domain config fragment file
Add the initial IPC/RPMsg domain config fragment file. This will
form the foundation for adding all the RPMsg related module
configurations, to be used by the TI Linux integration kernel.

Also, enable the fully upstreamed OMAP HwSpinlock driver while
at this including the HwSpinlock core. There are currently no
direct users of the HwSpinlock IP, but the built-in is the
preferred choice for this driver.

Signed-off-by: Suman Anna <s-anna@ti.com>