media: i2c: ov490: Use the host_control api instead of direct access
The original register initialization method did not handle
the proper handshake as defined by OmniVision.
The existing 5ms delay appears to be a shot in the dark
which was not long enough in all cases.
This caused the programming of the number of data lanes
to be used as a hit or miss situation sometime resulting
in a corrupted capture image.

The host_control API (which handles the proper handshakes
when programming registers) waits for the OV490
to acknowledge that the request has been accepted.

This also ensure that further register programming will
take effect properly.

Signed-off-by: Benoit Parrot <bparrot@ti.com>