ARM: dts: Add support for DRA71-LCard (aka J6Entry LCard)
The DRA71 LCARD ('L'ow 'C'ost 'A'utomotive 'R'eference 'D'esign), is a
low cost DRA71x reference design.
This support targets the rev B of the design.

Unlike other DRA7x EVMs, the DRA71 LCard, doesn't support Ethernet using
the Ethernet controller of the DRA71x. Instead the primary Ethernet option
is to use Ethernet over USB.

Unlike other DRA7x EVMs, the console is not on ttyS0 but uses ttyS2.

DRA71-LCard offers 2 displays: HDMI and FPDLink. The FPDLink display is
not supported.
MMC1 is used primarily for wireless, although there is a mux driven by a
GPIO to use an SD-Card instead.

Signed-off-by: Faiz Abbas <faiz_abbas@ti.com>
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>