drm/omap: dispc6: Fix broken replication logic in c8_to_c12()
The newly implemented replication logic in dispc6.c's c8_to_c12()
conversion function was broken. The patch fixes it. The idea is to use
full color scale of c12 format when converting c8 format to it. The 8
msb are straight from the c8, but the 4 msb are replicated to 4 lsb of
c12.

Signed-off-by: Jyri Sarha <jsarha@ti.com>